requirements_txt.txt
# requirements.txt - Alternative installation method
# Use this if conda is not available

# PDF Processing
PyPDF2==3.0.1

# Google Cloud Services  
google-cloud-texttospeech==2.16.3
google-generativeai==0.3.2

# Audio Processing
pydub==0.25.1
mutagen==1.47.0

# Configuration & Templates
PyYAML==6.0.1
python-dotenv==1.0.0

# Utilities
requests==2.31.0
platformdirs==4.0.0

# Development (optional)
pytest==7.4.3